pdf fetchers; simplified paper fetchers as functions instead of too generalized abstraction
