+++ main.py
@@ -1,0 +1,12 @@
+"""
+
+Extract papers
+
+"""
+
+import sys
+
+from python.util import parse_arguments
+
+if __name__ == '__main__':
+    args = parse_arguments(sys.argv[1:])
